Move lightnode into legacy
