+++ CMakeLists.txt
@@ -51,8 +51,10 @@ if(NOT ONLY_CPP_SDK)
     find_package(ittapi CONFIG REQUIRED)
 endif ()
 
+add_subdirectory(legacy)
+
 if(WITH_LIGHTNODE)
-    add_subdirectory(lightnode)
+    add_subdirectory(legacy/lightnode)
 endif()
 
 if(FULLNODE)
