+++ bcos-ledger/test/unittests/ledger/LedgerImplTest.cpp
@@ -1,4 +1,4 @@
-#include "bcos-ledger/LedgerImpl.h"
+#include <legacy/bcos-ledger/LedgerImpl.h>
 #include <bcos-concepts/ByteBuffer.h>
 #include <bcos-concepts/Serialize.h>
 #include <bcos-concepts/ledger/Ledger.h>
