+++ legacy/bcos-ledger/LedgerImpl.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Ledger.h"
+#include <bcos-ledger/Ledger.h>
 #include "bcos-framework/ledger/Ledger.h"
 #include "bcos-task/Task.h"
 #include <bcos-concepts/Basic.h>
@@ -27,15 +27,6 @@
 
 namespace bcos::ledger
 {
-static constexpr const int LIGHTNODE_MAX_REQUEST_BLOCKS_COUNT = 50;
-
-DERIVE_BCOS_EXCEPTION(NotFoundTransaction);
-DERIVE_BCOS_EXCEPTION(UnexpectedRowIndex);
-DERIVE_BCOS_EXCEPTION(MismatchTransactionCount);
-DERIVE_BCOS_EXCEPTION(MismatchParentHash);
-DERIVE_BCOS_EXCEPTION(NotFoundBlockHeader);
-DERIVE_BCOS_EXCEPTION(GetABIError);
-DERIVE_BCOS_EXCEPTION(GetBlockDataError);
 
 template <bcos::crypto::hasher::Hasher Hasher, bcos::concepts::storage::Storage Storage>
 class LedgerImpl : public bcos::concepts::ledger::LedgerBase<LedgerImpl<Hasher, Storage>>,
@@ -53,10 +44,10 @@ public:
         m_storage{std::move(storage)},
         m_merkle{m_hasher.clone()}
     {}
-    using statusInfoType = std::map<crypto::NodeIDPtr, bcos::protocol::BlockNumber>;
 
     void setKeyPageSize(size_t keyPageSize) { m_keyPageSize = keyPageSize; }
 
+protected:
     template <bcos::concepts::block::Block BlockType>
     void checkParentBlock(BlockType parentBlock, BlockType block)
     {
@@ -73,26 +64,6 @@ public:
             BOOST_THROW_EXCEPTION(MismatchParentHash{} << errinfo_comment{"No match parentHash!"});
         }
     }
-
-    crypto::NodeIDs filterSyncNodeList(
-        statusInfoType const& peersStatusInfo, bcos::protocol::BlockNumber needBlockNumber)
-    {
-        crypto::NodeIDs requestNodeIDList;
-        for (const auto& nodeStatus : peersStatusInfo)
-        {
-            LEDGER_LOG(INFO) << LOG_KV("nodeID", nodeStatus.first->hex())
-                             << LOG_KV("blockNumber: ", nodeStatus.second)
-                             << LOG_KV("，needBlockNumber: ", needBlockNumber);
-            if (nodeStatus.second >= needBlockNumber)
-            {
-                requestNodeIDList.push_back(nodeStatus.first);
-            }
-        }
-        LEDGER_LOG(DEBUG) << LOG_KV("requestNodeIDList size", requestNodeIDList.size());
-        return requestNodeIDList;
-    }
-
-private:
     template <bcos::concepts::ledger::DataFlag... Flags>
     task::Task<void> impl_getBlock(bcos::concepts::block::BlockNumber auto blockNumber,
         bcos::concepts::block::Block auto& block)
@@ -102,24 +73,6 @@ private:
         auto blockNumberStr = boost::lexical_cast<std::string>(blockNumber);
         (co_await getBlockData<Flags>(blockNumberStr, block), ...);
     }
-    template <bcos::concepts::ledger::DataFlag... Flags>
-    task::Task<void> impl_getBlockByNodeList(bcos::concepts::block::BlockNumber auto blockNumber,
-        bcos::concepts::block::Block auto& block, bcos::crypto::NodeIDs const& nodeList)
-    {
-        try
-        {
-            LEDGER_LOG(INFO) << "getBlockByNodeList: " << blockNumber;
-            auto blockNumberStr = boost::lexical_cast<std::string>(blockNumber);
-            (co_await getBlockData<Flags>(blockNumberStr, block), ...);
-        }
-        catch (NotFoundBlockHeader& e)
-        {
-            LEDGER_LOG(ERROR) << "Not found block";
-            block = {};
-        }
-        co_return;
-    }
-
     template <bcos::concepts::ledger::DataFlag... Flags>
     task::Task<void> impl_setBlock(bcos::concepts::block::Block auto block)
     {
@@ -299,85 +252,6 @@ private:
                           << status.blockNumber;
 
         co_return status;
-    }
-
-    task::Task<std::map<crypto::NodeIDPtr, bcos::protocol::BlockNumber>> impl_getAllPeersStatus()
-    {
-        std::map<crypto::NodeIDPtr, bcos::protocol::BlockNumber> allPeersStatus;
-        // assert(false); //never reach here
-        co_return allPeersStatus;
-    }
-
-    template <bcos::concepts::ledger::Ledger LedgerType, bcos::concepts::block::Block BlockType>
-    task::Task<size_t> impl_sync(LedgerType& source, bool onlyHeader)
-    {
-        auto& sourceLedger = bcos::concepts::getRef(source);
-        auto status = co_await impl_getStatus();
-        auto allPeersStatus = co_await sourceLedger.getAllPeersStatus();
-        bcos::protocol::BlockNumber currentMaxBlockNumber = 0;
-        for (const auto& nodeStatus : allPeersStatus)
-        {
-            if (nodeStatus.second > currentMaxBlockNumber)
-            {
-                currentMaxBlockNumber = nodeStatus.second;
-            }
-        }
-        LEDGER_LOG(DEBUG) << LOG_KV("allPeersStatus", allPeersStatus.size())
-                          << LOG_KV("currentMaxBlockNumber", currentMaxBlockNumber);
-        std::optional<BlockType> parentBlock;
-        size_t syncedBlock = 0;
-        auto syncBlockNumber = status.blockNumber + LIGHTNODE_MAX_REQUEST_BLOCKS_COUNT;
-        if (allPeersStatus.size() != 0 && currentMaxBlockNumber < syncBlockNumber)
-        {
-            syncBlockNumber = currentMaxBlockNumber;
-        }
-        // sync block
-        for (auto blockNumber = status.blockNumber + 1; blockNumber <= syncBlockNumber;
-            ++blockNumber)
-        {
-            LEDGER_LOG(INFO) << "Syncing block from remote: " << blockNumber << " | "
-                             << syncBlockNumber << " | " << onlyHeader;
-            BlockType block;
-            auto syncNodeList = filterSyncNodeList(allPeersStatus, blockNumber);
-            if (onlyHeader)
-            {
-                co_await sourceLedger.template getBlockByNodeList<bcos::concepts::ledger::HEADER>(
-                    blockNumber, block, syncNodeList);
-            }
-            else
-            {
-                co_await sourceLedger.template getBlockByNodeList<bcos::concepts::ledger::ALL>(
-                    blockNumber, block, syncNodeList);
-            }
-            // if getBlockByNodeList return empty block, break
-            if (::ranges::empty(block.blockHeader.data.parentInfo))
-            {
-                LEDGER_LOG(WARNING)
-                    << LOG_DESC("No blockHeader in block") << LOG_KV("blockNumber", blockNumber);
-                break;
-            }
-            if (blockNumber > 0)
-            {
-                if (!parentBlock)
-                {
-                    parentBlock = BlockType();
-                    co_await impl_getBlock<bcos::concepts::ledger::HEADER>(
-                        blockNumber - 1, *parentBlock);
-                }
-                checkParentBlock(*parentBlock, block);
-            }
-            if (onlyHeader)
-            {
-                co_await impl_setBlock<bcos::concepts::ledger::HEADER>(block);
-            }
-            else
-            {
-                co_await impl_setBlock<bcos::concepts::ledger::ALL>(block);
-            }
-            parentBlock = std::move(block);
-            ++syncedBlock;
-        }
-        co_return syncedBlock;
     }
 
     template <std::same_as<bcos::concepts::ledger::HEADER>>
