+++ legacy/concepts/bcos-concepts/scheduler/Scheduler.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "../Basic.h"
-#include "../protocol/Receipt.h"
-#include "../protocol/Transaction.h"
+#include <bcos-concepts/Basic.h>
+#include <bcos-concepts/protocol/Receipt.h>
+#include <bcos-concepts/protocol/Transaction.h>
 
 namespace bcos::concepts::scheduler
 {
