+++ legacy/lightnode/CMakeLists.txt
@@ -2,7 +2,7 @@ project(lightnode VERSION ${VERSION})
 
 add_library(bcos-lightnode INTERFACE)
 target_include_directories(bcos-lightnode INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
-target_link_libraries(bcos-lightnode INTERFACE ledger bcos-framework bcos-crypto bcos-task bcos-concepts)
+target_link_libraries(bcos-lightnode INTERFACE ledger bcos-framework bcos-crypto bcos-task bcos-concepts bcos-legacy)
 
 add_subdirectory(fisco-bcos-lightnode)
