+++ legacy/lightnode/bcos-lightnode/transaction-pool/TransactionPoolImpl.h
@@ -5,7 +5,7 @@
 #include "bcos-tars-protocol/protocol/TransactionImpl.h"
 #include "bcos-tars-protocol/protocol/TransactionReceiptImpl.h"
 #include "bcos-tars-protocol/tars/TransactionReceipt.h"
-#include <bcos-concepts/transaction-pool/TransactionPool.h>
+#include "../../../concepts/bcos-concepts/transaction-pool/TransactionPool.h"
 #include <bcos-framework/protocol/TransactionSubmitResult.h>
 #include <bcos-task/Task.h>
 #include <bcos-task/Wait.h>
