+++ legacy/lightnode/fisco-bcos-lightnode/CMakeLists.txt
@@ -1,5 +1,5 @@
 
-file(GLOB_RECURSE SOURCES "bcos-lightnode/*.h")
+file(GLOB_RECURSE SOURCES "*.h" "client/*.h")
 
 add_executable(fisco-bcos-lightnode ${SOURCES} main.cpp)
 target_link_libraries(fisco-bcos-lightnode PUBLIC bcos-cpp-sdk protocol_init gateway
