+++ legacy/lightnode/fisco-bcos-lightnode/client/LedgerClientImpl.h
@@ -8,7 +8,7 @@
 #include <bcos-concepts/ledger/Ledger.h>
 #include <bcos-framework/protocol/Protocol.h>
 #include <bcos-gateway/Gateway.h>
-#include <bcos-lightnode/Log.h>
+#include "../../bcos-lightnode/Log.h"
 #include <bcos-tars-protocol/tars/LightNode.h>
 #include <boost/algorithm/hex.hpp>
 #include <boost/throw_exception.hpp>
