+++ libinitializer/CMakeLists.txt
@@ -50,7 +50,7 @@ if(WITH_LEDGER_ELECTION)
 endif()
 if(WITH_LIGHTNODE)
     target_sources(init PRIVATE LightNodeInitializer.cpp LightNodeInitializer.h)
-    target_link_libraries(init PUBLIC bcos-lightnode bcos-concepts)
+    target_link_libraries(init PUBLIC bcos-lightnode bcos-concepts bcos-legacy)
 endif()
 set_target_properties(init PROPERTIES UNITY_BUILD "ON")
 add_dependencies(init BuildInfo.h)
